Josh Updates 7-23-24

+++ !Client Based Scripts/JME/ComputerEase_Service_Restart.ps1
@@ -20,7 +20,7 @@ $ProcessList = ("mm", "mmr")
 
 foreach ($Process in $ProcessList)
 {
-    ifif(Get-Process -Name $process -ErrorAction SilentlyContinue)
+    if(Get-Process -Name $process -ErrorAction SilentlyContinue)
     {
         Stop-Process -Name $Process -Force
     }
